client/requirements.txt
python-dotenv>=1.0
pydantic>=2.0
qdrant-client>=1.7
fastembed>=0.2
redis>=5.0
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-exporter-otlp-proto-http>=1.20
